✝ The Day We Meet Again ✝ The day we meet again I'll be waiting there I'll be waiting there for you 'Cause the years have been so lonely Like a dog without a home It's dangerous when you find out You've been drinking on your own
The day we meet again
We will walk in peace
Through the garden down the road
Where the mist of time is lifting
See it rising in the air
Like the shadow I was chasing
When I looked, it wasn't there, oh, no

Just in case you're wondering
What was really on my mind
It wasn't what you took, my love
It's what you left behind

And just in case you're wondering
Will it really be the same?
Will it really be the same?
You know we're only living for the day me meet again

So hold, hold on and don't let go
Time, it heals, you know, I know

The day we meet again
I'll be waiting there
I'll be waiting there for you
'Cause the years have been so lonely
Like a dog without a home
It's dangerous when you find out
You've been drinking on your own

The day we meet again
We will walk in peace
Through the garden down the road
Where the mist of time is lifting
See it rising in the air
Like the shadow I was chasing
When I looked, oh, no, it wasn't there, oh, no
It wasn't there, it wasn't there

It wasn't there, oh, no; oh, no; oh, no