☯️✝️☯️ Believe It Or Not ☯️✝️☯️
Believe it or not, everyone
Have things that they hide
Believe it or not, everyone
Keep most things inside

Believe it or not, everyone
Believe in something above
Believe it or not, everyone
Need to feel loved

Feel loved
But we don't, and we won't
Until we figure out

Could someone deliver us
And send us some kind of sign?
So close to giving up
'Cause faith is so hard to find

But you don't, and you won't
Until we figure out

I've seen it a lot, every time
The world turns upside down
Believe it or not, most of us
Feel like we're losing ground

Believe it or not, everyone
Hate admitting fear
Believe it or not, most of us
Wanna know why we're here

Why we're here
But we don't, and we won't
Until we figure out

Could someone deliver us
And send us some kind of sign?
So close to giving up
'Cause faith is so hard to find

Someone deliver us
Just send us some kind of sign
So close to giving up
'Cause faith is so hard to find

But you don't, you won't
Until we figure out

Most of us have nothing to complain about
Most of us have things we could live without
Everyone need advice on how to get along
You don't, we won't until we figure out

Believe it or not, everyone (everyone, everyone)

Believe it or not, everyone
Have things that they hide
Believe it or not, everyone
Keep most things inside

Believe it or not, everyone
Believe in something above
Believe it or not, everyone
Need to feel loved

Feel loved
But we don't, and we won't
Until we figure out

Could someone deliver us
And send us some kind of sign?
So close to giving up
'Cause faith is so hard to find

Someone deliver us
Just send us some kind of sign
So close to giving up
'Cause faith is so hard to find

But you don't, and you won't
Until we figure out
You don't, and you won't
Until we figure out