🫂 So Long Sentiment 🫂
Under this crushing wait, I'm sinking
It never goes away, the thinking
Echoes in empty rooms decaying
Alone to face the rest of me

So long sentiment
It doesn't matter now

What the hell is wrong with me?
Why am I torturing myself?
Inhaling all these memories
Like a breath of fire sent from hell

Lead in my gut, not in my spine
I feel distracted all the time
Well lucky me I'm finally all alone
I'll miss you

So long Sentiment
It doesn't matter now

What the hell is wrong with me?
Why am I torturing myself?
Fixated on these memories
Like a prisoner inside a cell

With nothing useful to say and no one to listen to it
Filling the deep with the pain, I slowly sink into it
Consider questionable things to try to get me through it
I've tried to push it away but I always give into it
Long for the taste of the rain that finally helps subdue it
And washes this all away

Under this crushing wait, I'm sinking
It never goes away, the thinking
Echoes in empty rooms are saying
"Time to erase the rest of me"