What Kind Of God ✝️☪️🕉️☸️✡️☯️☦️
I rode to my homeland
And carried the memory
Over the valley below
And the heart of a nation
The war cry was blazing
And the loss of my brothers and sons
Were all in the name of civilization

The empty moon saw I was leaving
With pride and hope in my heart
The mountains talk of
Far away dreaming

So came the invasion
Across from a faraway place
Into the new world
With guns and new order
They gave us religion
Then took away all of our rights
And forced us to live
Behind their new border

The four winds blow
And bring new vision
Where children laugh without fear
The land will speak
The people will listen

But yesterday my life was fading
With every reason to die
Ten miles to the north
As I rode my red horse
In the dying sun
And then when I saw
The white soldiers
One hundred or more in a line
Kill for the land
Saying their law will stand
What kind of god do they see?

The fighting is over
No matter what fear they might bring
My country will live
The dust clouds are falling
I stood by the river
And breathing the air that is life
Within the black hills
My spiritual calling

What kind of god can this be?