🚶‍♂️🚶‍♀️🚶😇 Angels Walk With You 😇🚶🚶‍♀️🚶‍♂️
I only know what I'm feeling
Do you know it's true
'Cos I've searched for you
'Cos I believe in you

Why do we walk in shadows
While this world falls apart
'Cos I'd die for you
'Cos I know angels walk with you

In my hands I hold nothing
The more I have the less I use
If I'd gone down that road forever
I would not have found you

And a new world with a new morning
A new sunrise on a brand new day
A new life with a real meaning
A new sunrise on a brand new day
'Cos angels walk with you

If we sit and do nothing
Will the light surely fade?
Can I trust in you forever?
If you show me the way

A new world with a new morning
A new sunrise on a brand new day
A new life with a real meaning
A new sunrise on a brand new day
'Cos angels walk with you
Yeah, walk with you, with you
Yeah, walk with you
'Cos I'd die for you
'Cos I know angels walk with you

A new world with a new morning
A new sunrise on a brand new day
A new life with a real meaning
A new sunrise on a brand new day
A new world with a new morning
A new sunrise on a brand new day
A new life with a real meaning
A new sunrise on a brand new day
'Cos angels walk with you