💀 Martyr 💀 Take the angels away The only ones who knew already know my name Times re-known for a war To be the only one that ever knew the way
Don’t be a martyr, it will be alright
I can see you've run out...
But still you make the most of making the most

I don’t want nobody to haunt me
Every time I see your face in the night
I know it’s all the voice in my head.
Time, time, it’s all been time
I feel the call of all the things that I dread
I know it's all a voice in my head...

Take the anger away
So many come, so many seem to go that way
It’s on now...
Digging a hole in the soul
To see the confines of the mental overload
You melted it all down

I don’t want nobody to haunt me
Every time I see your face in the night
I know it’s all the voice in my head.

But still you make the most of making the most.
I can see you’ve run out
But still you make the most of making the most.