🗺️🤔 Map of the Problematique 🤔🗺️
Fear, and panic in the air
I want to be free
From desolation and despair
And I feel like everything I sow
Is being swept away
Well, I refuse to let you go

I can't get it right
Get it right
Since I met you

Loneliness be over
When will this loneliness be over?

Life will flash before my eyes
So scattered and lost
I want to touch the other side
And no one thinks they are to blame
Why can't we see?
That when we bleed, we bleed the same

I can't get it right
Get it right
Since I met you

Loneliness be over
When will this loneliness be over?
Loneliness be over
When will this loneliness be over?