👹💀🦅 Monster / Suicide / America 👹💀🦅
[Monster]

Once the religious, the hunted and weary
Chasing the promise of freedom and hope
Came to this country to build a new vision
Far from the reaches of kingdom and pope

Like good Christians, some would burn the witches
Later some got slaves to gather riches

But still from near and far to seek America
They came by thousands to court the wild
But she just patiently smiled and bore a child
To be their spirit and guiding light

And once the ties with the crown had been broken
Westward in saddle and wagon it went
And 'til the railroad linked ocean to ocean
Many the lives which had come to an end

While we bullied, stole and bought our a homeland
We began the slaughter of the red man

But still from near and far to seek America
They came by thousands to court the wild
But she just patiently smiled and bore a child
To be their spirit and guiding light

The blue and grey they stomped it
They kicked it just like a dog
And when the war was over
They stuffed it just like a hog

And though the past has it's share of injustice
Kind was the spirit in many a way
But it's protectors and friends have been sleeping
Now it's a monster and will not obey

[Suicide]

The spirit was freedom and justice
And it's keepers seemed generous and kind
It's leaders were supposed to serve the country
But now they won't, pay it no mind
'Cause the people grew fat and got lazy
Now their vote is a meaningless joke
They babble about law and order
But it's all just an echo of what they've been told

Yeah, there's a monster on the loose
It's got our heads into the noose
And it just sits there watching

The cities have turned into jungles
And corruption is strangling the land
The police force is watching the people
And the people just can't understand
We don't know how to mind our own business
'Cause the whole world's got to be just like us
Now we are fighting a war over there
No matter who's the winner, we can't pay the cost

'Cause there's a monster on the loose
It's got our heads into the noose
And it just sits there, watching

[America]

America, where are you now?
Don't you care about your sons and daughters?
Don't you know, we need you now
We can't fight alone against the monster